requirements.txt
coloredlogs==15.0.1
humanfriendly==10.0
MouseInfo==0.1.3
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
pypiwin32==223
pyreadline3==3.4.1
PyRect==0.1.4
PyScreeze==0.1.28
python-dotenv==0.19.2
pytweening==1.0.4
pywin32==303
six==1.16.0
